+++ src/usecases/interfaces/ICrudRepository.ts
@@ -1,3 +1,4 @@
+import {DeepPartial} from '@steroidsjs/typeorm';
 import {SearchResultDto} from '../dtos/SearchResultDto';
 import {SearchInputDto} from '../dtos/SearchInputDto';
 import SearchQuery, {ISearchQueryConfig} from '../base/SearchQuery';
@@ -12,8 +13,19 @@ export interface ICrudRepository<TModel> {
     create: (model: TModel, transactionHandler?: TransactionHandler<TModel>) => Promise<TModel>,
     createQuery: (config?: ISearchQueryConfig<TModel>) => SearchQuery<TModel>,
     remove: (id: number, transactionHandler?: (callback) => Promise<void>) => Promise<void>,
-    save: (model: TModel, transactionHandler?: TransactionHandler<TModel>) => Promise<TModel>,
+    save(
+        model: TModel,
+        transactionHandler?: TransactionHandler<TModel | DeepPartial<TModel>>,
+    ): Promise<TModel>,
+    save(
+        model: DeepPartial<TModel>,
+        transactionHandler?: TransactionHandler<TModel | DeepPartial<TModel>>,
+    ): Promise<DeepPartial<TModel>>,
     softRemove: (id: number, transactionHandler?: (callback: () => Promise<void>) => Promise<void>) => Promise<void>,
-    update: (id: number, model: TModel, transactionHandler?: TransactionHandler<TModel>) => Promise<TModel>
+    update: (
+        id: number,
+        model: DeepPartial<TModel>,
+        transactionHandler?: TransactionHandler<TModel | DeepPartial<TModel>>,
+    ) => Promise<DeepPartial<TModel>>,
     isExistsById: (id: number) => Promise<boolean>,
 }