+++ src/usecases/interfaces/ISaveManager.ts
@@ -1,3 +1,5 @@
-export interface ISaveManager<TModel>{
-    save: (model) => Promise<TModel>,
+import {DeepPartial} from '@steroidsjs/typeorm';
+
+export interface ISaveManager<TModel> {
+    save: (model: TModel | DeepPartial<TModel>) => Promise<TModel | DeepPartial<TModel>>,
 }